feat(tests): suite de tests API complète — 68/68 tests
- Ajoute tests/api/*.test.ts couvrant les 9 zones : health, auth,
  quiz, classes, tests formels, admin-ped, directeur, parent, magic-links
- Ajoute tests/setup.ts (helpers api/login/register/uniqueEmail, auto-unwrap {ok,data})
- Ajoute scripts/seed-test-data.ts pour préparer la base de test
- Corrige ROUND(float, int) → ROUND(float::numeric, int) dans 6 routes
  (directeur classes, tests, élèves, évolution ; parent/summary ; results/classe/eleve)
- Corrige colonne t.title manquante : ALTER TABLE tests ADD COLUMN IF NOT EXISTS title TEXT
  + ajout dans db/schema.sql
- Corrige colonne a.time_spent_seconds inexistante dans directeur/classes/[id]/tests/[test_id]
- Scripts : bun run seed:test | bun run test | bun run test:watch

+++ app/api/directeur/classes/[id]/route.ts
@@ -31,9 +31,9 @@ export async function GET(_req: NextRequest, { params }: { params: Promise<{ id:
     SELECT t.id, COALESCE(t.title, 'Test #' || t.id) AS title,
            t.question_count, t.created_at,
            COUNT(s.id) FILTER (WHERE s.submitted_at IS NOT NULL) AS nb_soumis,
-           ROUND(AVG(s.score::float / t.question_count * 20) FILTER (WHERE s.submitted_at IS NOT NULL), 1) AS moyenne,
-           ROUND(MIN(s.score::float / t.question_count * 20) FILTER (WHERE s.submitted_at IS NOT NULL), 1) AS min_score,
-           ROUND(MAX(s.score::float / t.question_count * 20) FILTER (WHERE s.submitted_at IS NOT NULL), 1) AS max_score
+           ROUND((AVG(s.score::float / t.question_count * 20) FILTER (WHERE s.submitted_at IS NOT NULL))::numeric, 1) AS moyenne,
+           ROUND((MIN(s.score::float / t.question_count * 20) FILTER (WHERE s.submitted_at IS NOT NULL))::numeric, 1) AS min_score,
+           ROUND((MAX(s.score::float / t.question_count * 20) FILTER (WHERE s.submitted_at IS NOT NULL))::numeric, 1) AS max_score
     FROM test_assignments ta
     JOIN tests t ON t.id = ta.test_id
     LEFT JOIN sessions s ON s.test_id = t.id
